исправил все методы, исходя из тестов postman

+++ src/main/java/ru/yandex/practicum/filmorate/serviceDB/BasedFilmService.java
@@ -4,12 +4,16 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.FilmNotFoundException;
 import ru.yandex.practicum.filmorate.exception.FilmValidationException;
+import ru.yandex.practicum.filmorate.exception.GenreNotFoundException;
+import ru.yandex.practicum.filmorate.exception.MpaNotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.Genre;
+import ru.yandex.practicum.filmorate.model.MPA;
 import ru.yandex.practicum.filmorate.repository.FilmRepository;
 
 
 import java.util.List;
-import java.util.Optional;
+
 
 @Service
 @RequiredArgsConstructor
@@ -18,10 +22,18 @@ public class BasedFilmService implements FilmServiceDB {
 
 
     @Override
-    public Optional<Film> getFilmById(int id) {
-        final Film f = filmRepository.getFilmById(id)
-                .orElseThrow(() -> new FilmNotFoundException("Фильм с таким " + id + " не найден"));
-        return Optional.ofNullable(f);
+    public Film getFilmById(int id) {
+        final Film f = filmRepository.getFilmById(id);
+        if (f == null) {
+            throw new FilmNotFoundException("Фильм с таким " + id + " не найден");
+        }
+        return filmRepository.getFilmById(id);
+    }
+
+    @Override
+    public Film getFilmByIdWithGenre(int id) {
+
+        return filmRepository.getFilmByIdWithGenre(id);
     }
 
     @Override
@@ -31,15 +43,32 @@ public class BasedFilmService implements FilmServiceDB {
 
     @Override
     public Film createFilm(Film film) {
+        final MPA mpa = film.getMpa();
+        if (mpa == null || mpa.getId() < 1 || mpa.getId() > 6) {
+            throw new MpaNotFoundException("MPA not found");
+        }
+
+        final List<Genre> genres = film.getGenres();
+        if (genres != null && !genres.isEmpty()) {
+            for (Genre genre : genres) {
+                if (genre.getId() < 1 || genre.getId() > 7) {
+                    throw new GenreNotFoundException("Genre not found");
+                }
+            }
+        }
+
         return filmRepository.createFilm(film);
+
     }
 
     @Override
     public Film changeFilm(Film film) {
-        final Film f = filmRepository.getFilmById(film.getId())
-                .orElseThrow(() -> new FilmNotFoundException("Фильм с таким " + film.getId() + " не найден"));
-        filmRepository.changeFilm(f);
-        return f;
+        final Film f = filmRepository.getFilmById(film.getId());
+        if (f == null) {
+            throw new FilmNotFoundException("Фильм с таким " + film.getId() + " не найден");
+        }
+
+        return filmRepository.changeFilm(film);
     }
 
     @Override
@@ -50,8 +79,10 @@ public class BasedFilmService implements FilmServiceDB {
         if (userId < 1) {
             throw new FilmValidationException("Некорректный id пользователя");
         }
-        final Film f = filmRepository.getFilmById(id)
-                .orElseThrow(() -> new FilmNotFoundException("Фильм с таким " + id + " не найден"));
+        final Film f = filmRepository.getFilmById(id);
+        if (f == null) {
+            throw new FilmNotFoundException("Фильм с таким " + id + " не найден");
+        }
         filmRepository.addLike(id, userId);
     }
 
@@ -63,8 +94,10 @@ public class BasedFilmService implements FilmServiceDB {
         if (userId < 1) {
             throw new FilmValidationException("Некорректный id пользователя");
         }
-        final Film f = filmRepository.getFilmById(id)
-                .orElseThrow(() -> new FilmNotFoundException("Фильм с таким " + id + " не найден"));
+        final Film f = filmRepository.getFilmById(id);
+        if (f == null) {
+            throw new FilmNotFoundException("Фильм с таким " + id + " не найден");
+        }
         filmRepository.deleteLike(id, userId);
     }
